Organize Python semantic model types
