+++ crates/djls-project/src/lib.rs
@@ -23,22 +23,22 @@ pub use models::ModelGraph;
 pub use models::ModelId;
 pub use models::compute_model_graph;
 pub use project::Project;
-pub use python::CandidateKind;
-pub use python::CandidateLocation;
 pub use python::FileModuleDerivation;
 pub use python::FileModuleDerivationStatus;
-pub use python::FileResolutionDetail;
-pub use python::FileUnresolvedReason;
+pub use python::FileModuleResolution;
+pub use python::FileModuleUnresolvedReason;
 pub use python::Interpreter;
 pub use python::InvalidModuleName;
+pub use python::ModuleCandidate;
+pub use python::ModuleCandidateKind;
+pub use python::ModuleResolution;
+pub use python::ModuleUnresolvedReason;
 pub use python::PackageDirs;
 pub use python::PythonModule;
 pub use python::PythonModuleName;
-pub use python::ResolutionDetail;
 pub use python::ResolvedPrefix;
 pub use python::SearchPath;
 pub use python::SearchPaths;
-pub use python::UnresolvedReason;
 pub use python::file_to_module;
 pub use python::file_to_module_detail;
 pub use python::resolve_module_detail;
