+++ crates/djls-project/src/python.rs
@@ -20,18 +20,18 @@ pub(crate) use imports::ModuleKind;
 pub(crate) use imports::extract_import_bindings_for_source;
 pub(crate) use imports::import_bindings;
 pub use interpreter::Interpreter;
-pub use module::CandidateKind;
-pub use module::CandidateLocation;
 pub use module::FileModuleDerivation;
 pub use module::FileModuleDerivationStatus;
-pub use module::FileResolutionDetail;
-pub use module::FileUnresolvedReason;
+pub use module::FileModuleResolution;
+pub use module::FileModuleUnresolvedReason;
+pub use module::ModuleCandidate;
+pub use module::ModuleCandidateKind;
+pub use module::ModuleResolution;
+pub use module::ModuleUnresolvedReason;
 pub use module::PackageDirs;
-pub(crate) use module::PythonImport;
+pub(crate) use module::PythonImportRequest;
 pub use module::PythonModule;
-pub use module::ResolutionDetail;
 pub use module::ResolvedPrefix;
-pub use module::UnresolvedReason;
 pub use module::file_to_module;
 pub use module::file_to_module_detail;
 pub use module::resolve_module_detail;
