+++ crates/djls-project/src/python/semantic_model/evaluator.rs
@@ -9,12 +9,7 @@ use super::control_flow::BranchPath;
 use super::control_flow::Truthiness;
 use super::control_flow::evaluate_test_with;
 use super::control_flow::is_irrefutable_match_case;
-use super::model::PythonCompleteness;
-use super::model::PythonDict;
-use super::model::PythonDictEntry;
 use super::model::PythonSemanticModel;
-use super::model::PythonValue;
-use super::model::PythonValueKind;
 use super::mutation_target::MutationTarget;
 use super::source_graph::PythonImportEdge;
 use super::source_graph::PythonSourceGraph;
@@ -27,6 +22,11 @@ use super::touched_names::expr_read_names;
 use super::touched_names::first_import_segment;
 use super::touched_names::pattern_bound_names;
 use super::touched_names::target_write_names;
+use super::values::PythonCompleteness;
+use super::values::PythonDict;
+use super::values::PythonDictEntry;
+use super::values::PythonValue;
+use super::values::PythonValueKind;
 use crate::ast::ExprExt;
 use crate::ast::RangedExt;
 use crate::python::PythonSource;
